+++ src/twitch/handler/SettingsHandler.hpp
@@ -21,4 +21,5 @@ namespace SettingsHandler
     void handleGravitySettings(CommandSettingsPopup *popup, cocos2d::CCObject *sender);
     void handleSpeedSettings(CommandSettingsPopup *popup, cocos2d::CCObject *sender);
     void handleJumpscareSettings(CommandSettingsPopup *popup, cocos2d::CCObject *sender);
+    void handlePlayerEffectSettings(CommandSettingsPopup *popup, cocos2d::CCObject *sender);
 }